imdempotence for Poco build
based on folder as I didn't found any relevant file or folder to test

+++ officeonline-install.sh
@@ -80,11 +80,14 @@ if [ ! -d $poco ]; then
 fi
 
 ######## Poco Build ########
-cd "$poco"
-sudo -u lool ./configure | tee -a $log_file
-sudo -u lool make -j${cpu} | tee -a $log_file
-make install | tee -a $log_file
-
+## test if the poco poco has already been compiled
+# (the dir size should be around 450000ko vs 65000ko when no compilation)
+if [ $(du -s ${poco} | awk '{print $1}') -gt 100000 ]; then
+  cd "$poco"
+  sudo -u lool ./configure | tee -a $log_file
+  sudo -u lool make -j${cpu} | tee -a $log_file
+  make install | tee -a $log_file
+fi
 ###############################################################################
 ######## loolwsd Build ########
 #### Download dependencies ####
